# Repository.openapi+json.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "Repository API",
        "description": "",
        "version": "1.0"
    },
    "paths": {
...
        "/v1/maps/{mapId}": {
            "get": {
                "tags": [
                    "Maps"
                ],
                "parameters": [
                    {
                        "name": "mapId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Maps"
                ],
                "parameters": [
                    {
                        "name": "mapId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/maps/{gameType}/{mapName}": {
            "get": {
                "tags": [
                    "Maps"
                ],
                "parameters": [
                    {
                        "name": "gameType",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/GameType"
                        }
                    },
                    {
                        "name": "mapName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/maps": {
            "get": {
                "tags": [
                    "Maps"
                ],
                "parameters": [
                    {
                        "name": "gameType",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/GameType"
                        }
                    },
                    {
                        "name": "mapNames",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/MapsFilter"
                        }
                    },
                    {
                        "name": "filterString",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "skipEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "takeEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/MapsOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Maps"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "Maps"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/v1/maps/{mapId}/image": {
            "post": {
                "tags": [
                    "Maps"
                ],
                "parameters": [
                    {
                        "name": "mapId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GameType": {
                "enum": [
                    "Unknown",
                    "CallOfDuty2",
                    "CallOfDuty4",
                    "CallOfDuty5",
                    "Insurgency",
                    "ArkSurvivalEvolved",
                    "Battlefield1",
                    "Battlefield3",
                    "Battlefield4",
                    "Battlefield5",
                    "BattlefieldBadCompany2",
                    "CrysisWars",
                    "Left4Dead2",
                    "Minecraft",
                    "PlayerUnknownsBattleground",
                    "RisingStormVietnam",
                    "Rust",
                    "WarThunder",
                    "WorldOfWarships",
                    "WorldWar3",
                    "UnrealTournament2004",
                    "Arma",
                    "Arma2",
                    "Arma3"
                ],
                "type": "string",
                "format": "int32"
            },
...
            "MapsFilter": {
                "enum": [
                    "EmptyMapImage"
                ],
                "type": "string",
                "format": "int32"
            },
            "MapsOrder": {
                "enum": [
                    "MapNameAsc",
                    "MapNameDesc",
                    "GameTypeAsc",
                    "GameTypeDesc",
                    "PopularityAsc",
                    "PopularityDesc"
                ],
                "type": "string",
                "format": "int32"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}